Break up asset part views into common traits and split up the classes
