+++ app/src/main/res/layout/message_video_asset_content.xml
@@ -25,7 +25,7 @@
              android:layout_height="match_parent"
     >
 
-    <com.waz.zclient.messages.parts.AssetActionButton
+    <com.waz.zclient.views.AssetActionButtonNew
         android:id="@+id/action_button"
         android:layout_width="@dimen/content__video_message__button__size"
         android:layout_height="@dimen/content__video_message__button__size"
