+++ app/src/main/scala/com/waz/zclient/WireContext.scala
@@ -20,8 +20,6 @@ package com.waz.zclient
 import android.annotation.SuppressLint
 import android.app.{Activity, Service}
 import android.content.{Context, ContextWrapper}
-import android.graphics.drawable.Drawable
-import android.graphics.{ColorFilter, PixelFormat}
 import android.support.v4.app.{Fragment, FragmentActivity}
 import android.view.{LayoutInflater, View, ViewGroup, ViewStub}
 import com.waz.ZLog
@@ -111,21 +109,6 @@ object ViewHelper {
     case vg: ViewGroup => vg
     case _ => view.getParent.asInstanceOf[ViewGroup]
   }
-}
-
-trait DrawableHelper extends Drawable with Drawable.Callback with Injectable {
-
-  override def setColorFilter(colorFilter: ColorFilter): Unit = ()
-
-  override def setAlpha(alpha: Int): Unit = ()
-
-  override def getOpacity: Int = PixelFormat.TRANSLUCENT
-
-  override def scheduleDrawable(who: Drawable, what: Runnable, when: Long): Unit = scheduleSelf(what, when)
-
-  override def invalidateDrawable(who: Drawable): Unit = invalidateSelf()
-
-  override def unscheduleDrawable(who: Drawable, what: Runnable): Unit = unscheduleSelf(what)
 }
 
 trait ServiceHelper extends Service with Injectable with WireContext with EventContext {
